srcalc/src/test/java/gov/va/med/srcalc/domain/CalculationTest.java
package gov.va.med.srcalc.domain;

import static org.junit.Assert.*;

import org.joda.time.DateTime;
import org.junit.Test;

public class CalculationTest
{
    protected Patient dummyPatient()
    {
        return new Patient(1, "Zach Smith");
    }

    @Test
    public final void testForPatient()
    {
        final Patient patient = dummyPatient();
        final DateTime testStartDateTime = new DateTime();
        final Calculation c = Calculation.forPatient(patient);
        
        assertTrue("start date not in the past",
                // DateTime has millisecond precision, so the current time may
                // still be the same. Use "less than or equal to".
                c.getStartDateTime().compareTo(new DateTime()) <= 0);
        assertTrue("start date not after test start",
                c.getStartDateTime().compareTo(testStartDateTime) >= 0);
        assertEquals(patient, c.getPatient());
    }
    
    @Test(expected = IllegalStateException.class)
    public final void testGetVariablesIllegal()
    {
        Calculation.forPatient(dummyPatient()).getVariables();
    }
    
}
